% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditionManipulateCompactiffy.R
\name{condCompactify}
\alias{condCompactify}
\title{This function takes a list of moment-condition defining data frames and a vector of corresponding coefficients, equal in length to the list. The data frame list is searched for identical entries, which are made unique, and the corresponding coefficients are summed.}
\usage{
condCompactify(cond.list, cond.coeff, cond.linTrans)
}
\arguments{
\item{cond.list}{A list of data frames defining elements of the moment condition.}

\item{cond.coeff}{A list of coefficients for taking a linear combination of the individual moment conditions defined by \code{cond.list}; must be of equal length as \code{cond.list}}
}
\value{
List with named elements, \code{cond.list} is a list akin to the input list, \code{cond.coeff} is a list of corresponding coefficients.
}
\description{
This function takes a list of moment-condition defining data frames and a vector of corresponding coefficients, equal in length to the list. The data frame list is searched for identical entries, which are made unique, and the corresponding coefficients are summed.
}

