% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditionGeneratePortfolio.R
\name{condGeneratePortfolio}
\alias{condGeneratePortfolio}
\title{Prepare elements for hedged GLT return moment calculation}
\usage{
condGeneratePortfolio(u.vec, tau.js = 1/2 * 5/252, deltas = 1/252/6.5,
  ttm = 1/12, weights = rep(1, length(tau.js)), N.factors = 3)
}
\arguments{
\item{u.vec}{The frequency vector for which the portfolio is built. Each frequency represents a cross-product of contemporeneous hedged returns.}

\item{tau.js}{Vector of horizons for which the return is calculated}

\item{deltas}{Vector of hedging frequencies; if \code{length(detlas) == 1}, \code{rep} is applied to have length equal to \code{length(tau.js)}}

\item{ttm}{Vector of initial maturity of the portfolio.}

\item{weights}{A vector of weights to pre-multiply the returns (useful for setting up quadrature approximations)}
}
\value{
return.df.list List with two fields: cond.list describing the moment conditions and cond.coeff with coefficients for linear combination.
}
\description{
Generate list of data frames that describes a sum of hedged GLT portfolio returns or their squares.
}

