% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionManipulateMultiply.R
\name{condMultiply}
\alias{condMultiply}
\title{This function takes two data frame lists which define elements of moment conditions, and two vectors of corresponding linear combination coefficients, and combines them into a single list of data frames which would result from multiplying the two expressions rational in GLT-portfolios.}
\usage{
condMultiply(condStruct.1, condStruct.2, lag = 0, compactify = FALSE,
  N.factors = 3)
}
\arguments{
\item{lag}{Positive real, the time distance between elements of \code{cond.list.1} and \code{cond.list.2}, the former happening ``later''.}

\item{cond.list.1}{A list of data frames defining elements of the moment condition.}

\item{cond.list.2}{A list of data frames defining elements of the moment condition.}

\item{coeffs.1}{A list of coefficients for taking a linear combination of the individual moment conditions defined by \code{cond.list.1}; must be of equal length as \code{cond.list.1}}

\item{coeffs.2}{A list of coefficients for taking a linear combination of the individual moment conditions defined by \code{cond.list.2}; must be of equal length as \code{cond.list.2}}
}
\value{
List with named elements, \code{cond.list} is a list akin to the input lists, \code{cond.coeff} is a list of corresponding coefficients.
}
\description{
This function takes two data frame lists which define elements of moment conditions, and two vectors of corresponding linear combination coefficients, and combines them into a single list of data frames which would result from multiplying the two expressions rational in GLT-portfolios.
}
