% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condRegularize.R
\name{condRegularize}
\alias{condRegularize}
\title{Ensure that condition lists have uniform columns}
\usage{
condRegularize(cond.struct, N.factors = 3)
}
\arguments{
\item{cond.struct}{The condition structure to be regularized}

\item{N.factors}{The number of volatility factors assumed in the calculations.}
}
\value{
Returns a regularized condition list.
}
\description{
This function takes a condition list and fills it up the missing columns with 0s.
}

