% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionStockChange.R
\name{condStockChange}
\alias{condStockChange}
\title{Return condition list for stock return}
\usage{
condStockChange(ret.period = 5/252, ret.start = 0, N.factors = 3)
}
\arguments{
\item{ret.period}{The period over which we are interested in the change}
}
\value{
A standard cond struct list
}
\description{
Return condition list for stock return
}
