% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionVolChange.R
\name{condVolChange}
\alias{condVolChange}
\title{Return condition list for approximate volatility change}
\usage{
condVolChange(factorNum = 1, ret.period = 5/252, ret.start = 0,
  lin.u = 0.001, N.factors = 3)
}
\arguments{
\item{factorNum}{Integer, specifying which factor we are talking about}

\item{ret.period}{The period over which we are interested in the change}

\item{lin.u}{A small number such that expm(lin.u * V) is nearly equal to V}
}
\value{
A standard cond struct list
}
\description{
Return condition list for approximate volatility change
}

