\name{divergenceModelR-package}
\alias{divergenceModelR-package}
\alias{divergenceModelR}
\docType{package}
\title{
\packageTitle{divergenceModelR}
}
\description{
\packageDescription{divergenceModelR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{divergenceModelR}
\packageIndices{divergenceModelR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{divergenceModelR}

Maintainer: \packageMaintainer{divergenceModelR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
