% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hedgedReturn.R
\name{hedged.return}
\alias{hedged.return}
\title{Create condition structure for hedged option return}
\usage{
hedged.return(u, tau.j, delta, ttm, N.factors = 3)
}
\arguments{
\item{u}{The frequency corresponding to the option portfolio}

\item{tau.j}{Start of the return period}

\item{delta}{Length of the return period}

\item{ttm}{The time-to-maturity of the option at initiation}
}
\value{
A usual structure list
}
\description{
Create condition structure for hedged option return
}

