% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrVarRP.R
\name{integrVarRP}
\alias{integrVarRP}
\title{Integrated variance risk premia}
\usage{
integrVarRP(params.structure, v.0 = NULL, tT = c(0, 1), grid.d = 1000,
  N.factors = 3)
}
\arguments{
\item{params.structure}{a named list of named lists that describe the \code{P} and \code{Q} behaviour of the stochastic volatility model.}

\item{v.0}{initial condition for \code{IVRP} calculation, if \code{NULL}, long-run factor mean values are used instead.}

\item{tT}{vector of length at least 2, start- and end-point of calculation. If the vector is lonver, the VRP will be returned at the intermediate values as well. Otherwise, the output values are determined via \code{rp.control$grid.d}.}

\item{grid.d}{determines the number of intervals into which the \code{tT} period is divided}
}
\value{
Returns a list with fields \code{tot}, \code{cont} and \code{jmp}, referring to, respectively, the total risk premium, its continuous and discontinuous parts.
}
\description{
Given the parameters and initial values, this function produces the term structure of the integrated variance risk premium. Definition of IVRP as in reference below. This function treats the case of no volatility jumps if \code{control$use.CIR = TRUE}.
}
\references{
Ait-Sahalia, Y. and Karaman, M., and Mancini, L., ``The Term Structure of Variance Risk Premia and the Expectation Hypothesis''
}

