% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanAndcovList.R
\name{modelDynamics}
\alias{modelDynamics}
\title{Jump-diffusion model dynamics}
\usage{
modelDynamics(params.P, params.Q, dT = 5/252, rtol = 1e-14, N.factors = 3,
  N.points = 2,
  jumpTransform = getPointerToJumpTransform("expNormJumpTransform")$TF,
  mod.type = "standard", ...)
}
\arguments{
\item{params.P}{parameter list}

\item{params.Q}{parameter list risk-neutral}

\item{dT}{The Time horizon}

\item{rtol}{The relative tolerance required in the integration}

\item{N.factors}{The number of stochastic vol factors assumed.}

\item{N.points}{Number of points in sparse grid quadratures. 2 usually works pretty good, more than 3 is overkill.}

\item{jumpTransform}{pointer to jump trasform c++ function}

\item{mod.type}{character, \code{'standard'} or \code{'casc.vol'}}
}
\value{
Returns 2 lists and a matrix.
}
\description{
This function creates mean and covariance structures that can be passed to linCombMean to combine into an estimate of the conditional first and second (cross-)moments of the stock return and volatility factors.
}
