% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelEvaluation.R
\name{modelEvaluation}
\alias{evaluation_VCOVsandwich}
\alias{evaluation_fittedPrices}
\alias{evaluation_plotting}
\alias{modelEvaluation}
\title{Affine model evaluation}
\usage{
evaluation_fittedPrices(par.vec, par.list = NULL, data.structure, model.spec,
  filterFoo = DSQ_sqrtFilter, handlerFoo = affineObservationStateHandler,
  N.points = 5)

evaluation_plotting(par.vec, par.list = NULL, data.structure, model.spec,
  filterFoo = DSQ_sqrtFilter, handlerFoo = affineObservationStateHandler,
  N.points = 5, plot.path)

evaluation_VCOVsandwich()
}
\arguments{
\item{data.structure}{\code{list} with fields \code{spec.mat}, \code{obs.data}, \code{noise.cov.cube}}

\item{model.spec}{\code{list} with fields \code{params.P}, \code{params.Q}, \code{jump.type}, \code{dt}, \code{N.factors}, \code{error}, \code{mkt}}
}
\value{
\code{evaluation_fittedPrices} returns a list with fields \code{obs} (the observed returns and swap prices) and \code{fit} (the predicted returns and fitted swap prices.)

\code{evaluation_plotting} does not return.
}
\description{
Functions for evaluating fit of models estimated on divergence data
}
\details{
\code{evaluation_plotting} plots the following: (1) time series of filtered factor values, (2) time series of fitted and observed swap rates, (3) Q-Q plots of filtered factors and bootstrap sample from the estimated model factors, (4) autocorrelation functions of pricing errors, (5)
}
\section{Functions}{
\itemize{
\item \code{evaluation_plotting}: 

\item \code{evaluation_VCOVsandwich}: 
}}

