% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/specMakers.R
\name{modSetup}
\alias{modSetup}
\alias{specData_DSQ_1M_6M_0115}
\alias{specData_DS_1M_6M_0115}
\alias{specData_D_1M_6M_0115}
\alias{spec_1FtoyModel}
\alias{spec_3FsepIntModel}
\title{Model setup -- specification generators}
\usage{
spec_1FtoyModel()

specData_DSQ_1M_6M_0115(path.to.data)

specData_DS_1M_6M_0115(path.to.data)

specData_D_1M_6M_0115(path.to.data)

spec_3FsepIntModel()
}
\value{
\code{model.spec} and \code{data.structure} objects
}
\description{
Functions which fix model and data specifications for estimation. These functions are not intended for use for the public. They might get separated from the package in the future and placed in a separate archive.
}
\details{
\code{spec_1FtoyModel} Specifies a single-factor model with leverage effect, exponential jumps in volatility and correlated double exponential jumps in the underlying, with same tail parameter on both sides.

\code{specData_DSQ_1M_6M_0115} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error varcov matrices.

\code{specData_DS_1M_6M_0115} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error varcov matrices. Excludes quarticity

\code{specData_D_1M_6M_0115} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error varcov matrices. Excludes skewness and quarticity

\code{spec_3FsepIntModel} Specifies a three-factor model with leverage effect, exponential jumps in volatility and correlated double exponential jumps in the underlying, with same tail parameter on both sides. One of the factors only drives jump intensity, not the continuous volatility
}
\section{Functions}{
\itemize{
\item \code{specData_DSQ_1M_6M_0115}: 

\item \code{specData_DS_1M_6M_0115}: 

\item \code{specData_D_1M_6M_0115}: 

\item \code{spec_3FsepIntModel}: 
}}

