% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIXfutures.R
\name{vixfut}
\alias{priceVIX}
\alias{priceVIXFUT}
\alias{vixfut}
\title{VIX index and VIX FUTURES pricing}
\usage{
priceVIX(v.0, t.vec, params.Q, ...)

priceVIXFUT(v.0, t, tau, params.Q, ...)
}
\arguments{
\item{v.0}{initial variance factor value, size \code{S x (N.factors+1)} for \code{priceVIX}, size \code{1 x (N.factors+1)} for \code{priceVIXFUT}.}

\item{t.vec}{maturity of VIX index, length \code{T}.}

\item{params.Q}{Q measure parameters.}

\item{...}{arguments to ODE solution routines}

\item{t}{maturity of futures contract, length 1.}

\item{tau}{maturity of VIX index, on which the futures are written, length 1.}
}
\value{
VIX futures value in 1-factor model with v.0 starting values, VIX value
}
\description{
Calculate the VIX index and the price of VIX futures contracts with arbitrary maturities in affine jump diffusion models.
}
\details{
\code{priceVIX} calculates the VIX index for an arbitrary set of starting factor values, dates and maturities. \code{priceVIXFUT} only works for a single index time to maturity and a single futures time to maturity. Time to maturity of the index, on which the futures are written, is given net of the futures maturity, i.e. for calculating the price of a futures contract, that matures in 6 months, on the index value between 6 months and 7 months from now, give \code{t = 6/12} and \code{tau = 1/12}.
}

