% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{portfolio_sqrtFilter}
\alias{portfolio_sqrtFilter}
\alias{filterCaller}
\alias{DSQ_sqrtFilter}
\alias{D_sqrtFilter}
\alias{DS_sqrtFilter}
\title{C++ filter call}
\usage{
portfolio_sqrtFilter(dataMat, initState, initProcCov, modelParams)

DSQ_sqrtFilter(dataMat, initState, initProcCov, modelParams)

D_sqrtFilter(dataMat, initState, initProcCov, modelParams)

DS_sqrtFilter(dataMat, initState, initProcCov, modelParams)
}
\description{
This function wraps the ukfClass construction and exposes the call to R. Should not be used directly by the user, rather its exported R interface should be given as an argument to \code{\link{modelLikelihood}}
}
\details{
\code{portfolio_sqrtFilter} works on whatever portfolios you supply

\code{DSQ_sqrtFilter} works on divergence, skewness and quarticity

\code{D_sqrtFilter} works on divergence

\code{DS_sqrtFilter} works on divergence
}
