% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specMakers.R
\name{modSetup}
\alias{modSetup}
\alias{specData_DSQ_1M_6M_0115_extraNoise}
\alias{specData_DS_1M_6M_0115_extraNoise}
\alias{specData_D_1M_6M_0115_extraNoise}
\alias{specData_PF_DSQ_3Mat}
\alias{spec_1FtoyModel_extraNoise}
\alias{spec_3FsepIntModel_erp_extraNoise}
\alias{spec_3FsepIntModel_erp_portfolio}
\alias{spec_3FsepIntModel_smpl}
\title{Model setup -- specification generators}
\usage{
specData_DSQ_1M_6M_0115_extraNoise(path.to.data)

specData_DS_1M_6M_0115_extraNoise(path.to.data)

specData_D_1M_6M_0115_extraNoise(path.to.data)

spec_1FtoyModel_extraNoise(U)

spec_3FsepIntModel_erp_extraNoise(U)

spec_3FsepIntModel_erp_portfolio()

spec_3FsepIntModel_smpl()

specData_PF_DSQ_3Mat(path.to.data)
}
\arguments{
\item{U}{number of observed pfolios}
}
\value{
\code{model.spec} and \code{data.structure} objects
}
\description{
Functions which fix model and data specifications for estimation. These functions are not intended for use for the public. They might get separated from the package in the future and placed in a separate archive.
}
\details{
\code{specData_DSQ_1M_6M_0115_extraNoise} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error correlation matrices.

\code{specData_DS_1M_6M_0115_extraNoise} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error correlation matrices. Excludes quarticity

\code{specData_D_1M_6M_0115_extraNoise} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error correlation matrices. Excludes skewness and quarticity

\code{spec_1FtoyModel_extraNoise}  Specifies a single-factor model with leverage effect, exponential jumps in volatility and correlated double exponential jumps in the underlying, with same tail parameter on both sides. There are variance parameters for observation noise.

\code{spec_3FsepIntModel_erp_extraNoise} Specifies a three-factor model with leverage effect, exponential jumps in volatility and correlated double exponential jumps in the underlying, with same tail parameter on both sides. One of the factors only drives jump intensity, not the continuous volatility. There are variance parameters for observation noise. Plus, there is a more flexible parametrisation of the equity risk premium parameters, with some Q parameters restricted. Finally, the first vol factor has pure-jump dynamics

\code{spec_3FsepIntModel_erp} Specifies a three-factor model with leverage effect, exponential jumps in volatility and correlated double exponential jumps in the underlying, with same tail parameter on both sides. One of the factors only drives jump intensity, not the continuous volatility. There is a more flexible parametrisation of the equity risk premium parameters, with some Q parameters restricted. Finally, the first vol factor has pure-jump dynamics

\code{specData_DS_1M_6M_0115_cor} returns loads the divergence sample from 2001 to 2015, p= 0.5, maturities 1/12 and 1/2 years. Includes SP500 returns from OMTR and the bootstrapped error correlation matrices. Excludes quarticity
}

